+++ .env.example
@@ -47,8 +47,8 @@ PAYPAL_DONATE_LINK=https://www.sandbox.paypal.com/cgi-bin/webscr?cmd=_s-xclick&h
 # EMAIL_IR=ir@example.com
 # EMAIL_PAYMENT=payment@example.com
 
-# PIWIK_URL_BASE=https://piwik.example.com
-# PIWIK_SITE_ID=1
+# MATOMO_URL_BASE=https://matomo.example.com
+# MATOMO_SITE_ID=1
 
 # styleguide theme
 # SG_COLORS=
