+++ lib/constants.js
@@ -20,8 +20,8 @@ export const RENDER_FRONTEND_BASE_URL =
 export const PUBLIKATOR_BASE_URL = ENV.PUBLIKATOR_BASE_URL
 export const ADMIN_BASE_URL = ENV.ADMIN_BASE_URL
 
-export const PIWIK_URL_BASE = ENV.PIWIK_URL_BASE
-export const PIWIK_SITE_ID = ENV.PIWIK_SITE_ID
+export const MATOMO_URL_BASE = ENV.MATOMO_URL_BASE
+export const MATOMO_SITE_ID = ENV.MATOMO_SITE_ID
 
 export const SG_COLORS = ENV.SG_COLORS
 export const SG_FONT_STYLES = ENV.SG_FONT_STYLES
