+++ lib/translations.json
@@ -1,5 +1,5 @@
 {
-  "updated": "2021-04-29T15:20:26.610Z",
+  "updated": "2021-05-05T11:24:57.913Z",
   "title": "live",
   "data": [
     {
@@ -4011,47 +4011,47 @@
       "value": "Inhalt nur für Mitglieder. <a href=\"/probelesen\">Probelesen?</a>"
     },
     {
-      "key": "piwik/optOut/label",
+      "key": "matomo/optOut/label",
       "value": "Ihr Tracking-Status"
     },
     {
-      "key": "piwik/optOut/noscript",
+      "key": "matomo/optOut/noscript",
       "value": "Aktivieren Sie JavaScript, um Ihren Status zu sehen."
     },
     {
-      "key": "piwik/optOut/status/loading",
+      "key": "matomo/optOut/status/loading",
       "value": "lädt…"
     },
     {
-      "key": "piwik/optOut/status/timeout",
+      "key": "matomo/optOut/status/timeout",
       "value": "Konnte nicht geladen werden. Bitte versuchen Sie es später nochmals."
     },
     {
-      "key": "piwik/optOut/status/fail",
+      "key": "matomo/optOut/status/fail",
       "value": "Konnte nicht gespeichert werden. Bitte versuchen Sie es später nochmals."
     },
     {
-      "key": "piwik/optOut/status/on",
+      "key": "matomo/optOut/status/on",
       "value": "Aktiv. Danke!"
     },
     {
-      "key": "piwik/optOut/status/off",
+      "key": "matomo/optOut/status/off",
       "value": "Sie haben das Tracking deaktiviert. Bitte beachten Sie, dass Sie, wenn Sie einen anderen Computer, das Deaktivierungs-Cookie löschen oder einen anderen Webbrowser verwenden, die Deaktivierungsprozedur nochmals absolvieren müssen."
     },
     {
-      "key": "piwik/optOut/status/saving",
+      "key": "matomo/optOut/status/saving",
       "value": "speichert…"
     },
     {
-      "key": "piwik/optOut/status/dnt",
+      "key": "matomo/optOut/status/dnt",
       "value": "Das Tracking ist bei Ihnen derzeit nicht aktiv, denn Ihr Browser hat uns mitgeteilt, dass Sie kein Tracking wünschen."
     },
     {
-      "key": "piwik/optOut/turn/off",
+      "key": "matomo/optOut/turn/off",
       "value": "Deaktivieren?"
     },
     {
-      "key": "piwik/optOut/turn/on",
+      "key": "matomo/optOut/turn/on",
       "value": "Aktivieren?"
     },
     {
